twitter-ai-bot-api/pyproject.toml
[tool.poetry]
name = "twitter-ai-bot-api"
version = "0.1.0"
description = ""
authors = ["saishftw <ksaish1998@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
langchain = "^0.3.8"
langchain-xai = "^0.1.0"
langchain-google-vertexai = "^2.0.7"
anthropic = "^0.39.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
